reorganize slack code
